requirements.txt
certifi==2023.7.22
chardet==5.1.0
idna==3.4
python-telegram-bot==20.3
requests==2.31.0
urllib3==2.0.7
yt-dlp==2023.7.6
